backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
chromadb
pypdf
sentence-transformers
python-multipart
numpy<2
